refactor, extract HTTPS server instantiation to module injector
